+++ docs/en/start.md
@@ -24,30 +24,30 @@ Each chart component of v-charts is individually packaged under the lib folder
 
 ```
 |- lib/
-    |- line.js  -------------- Line Chart
-    |- bar.js  --------------- Bar Chart
-    |- histogram.js  --------- Histogram Chart
-    |- pie.js  --------------- Pie Chart
-    |- ring.js  -------------- Ring Chart
-    |- funnel.js  ------------ Funnel Chart
-    |- waterfall.js  --------- Waterfall Chart
-    |- radar.js  ------------- Radar Chart
-    |- map.js  --------------- Map Chart
-    |- sankey.js  ------------ Sankey Chart
-    |- heatmap.js  ----------- Heatmap Chart
-    |- scatter.js  ----------- Scatter Chart
-    |- candle.js  ------------ Candle Chart
-    |- gauge.js  ------------- Gauge Chart
-    |- tree.js  -------------- Tree Chart
-    |- bmap.js  -------------- Baidu map
-    |- amap.js  -------------- Amap
+    |- line.common.js  -------------- Line Chart
+    |- bar.common.js  --------------- Bar Chart
+    |- histogram.common.js  --------- Histogram Chart
+    |- pie.common.js  --------------- Pie Chart
+    |- ring.common.js  -------------- Ring Chart
+    |- funnel.common.js  ------------ Funnel Chart
+    |- waterfall.common.js  --------- Waterfall Chart
+    |- radar.common.js  ------------- Radar Chart
+    |- map.common.js  --------------- Map Chart
+    |- sankey.common.js  ------------ Sankey Chart
+    |- heatmap.common.js  ----------- Heatmap Chart
+    |- scatter.common.js  ----------- Scatter Chart
+    |- candle.common.js  ------------ Candle Chart
+    |- gauge.common.js  ------------- Gauge Chart
+    |- tree.common.js  -------------- Tree Chart
+    |- bmap.common.js  -------------- Baidu map
+    |- amap.common.js  -------------- Amap
 ```
 
 When used, a single chart component can be directly import into the project.
 
 ```js
 import Vue from 'vue'
-import VeLine from 'v-charts/lib/line'
+import VeLine from 'v-charts/lib/line.common'
 import App from './App.vue'
 
 Vue.component(VeLine.name, VeLine)
