+++ docs/index.html
@@ -26,18 +26,20 @@
   <div id="app"></div>
   <script>
     window.VUEP_FORM_RESOURCE = [
-      '<script src="https://cdn.jsdelivr.net/npm/vue/dist/vue.min.js"><\/script>',
-      '<script src="https://cdn.jsdelivr.net/npm/echarts/dist/echarts.min.js"><\/script>',
-      '<script src="https://cdn.jsdelivr.net/npm/echarts-amap/dist/echarts-amap.min.js"><\/script>',
-      '<script src="https://cdn.jsdelivr.net/npm/echarts/dist/extension/bmap.min.js"><\/script>',
-      '<script src="https://cdn.jsdelivr.net/npm/v-charts/lib/index.min.js"><\/script>',
-      '<link rel="stylesheet" href="https://cdn.jsdelivr.net/npm/v-charts/lib/style.min.css">'
+      '<script src="https://cdn.jsdelivr.net/npm/v-charts/lib/index.min.js"><\/script>'
+    ]
+    window.VUEP_RESOURCE = [
+      "https://cdn.jsdelivr.net/npm/vue/dist/vue.min.js",
+      "https://cdn.jsdelivr.net/npm/echarts/dist/echarts.min.js",
+      "https://cdn.jsdelivr.net/npm/echarts-amap/dist/echarts-amap.min.js",
+      "https://cdn.jsdelivr.net/npm/echarts/dist/extension/bmap.min.js",
+      "https://cdn.jsdelivr.net/npm/v-charts/lib/style.min.css"
     ]
     window.$docsify = {
       name: 'v-charts',
       repo: 'https://github.com/elemefe/v-charts',
       loadSidebar: 'sidebar.md',
-      ga: 'UA-104697795-1',
+      ga: 'UA-120605025-3',
       homepage: 'install.md',
       formatUpdated: '{MM}/{DD} {HH}:{mm}',
       auto2top: true,
@@ -98,7 +100,7 @@
   <script src="//cdn.jsdelivr.net/npm/babel-standalone/babel.min.js"></script>
   <script src="//cdn.jsdelivr.net/npm/shap/sdk/Shap.umd.min.js"></script>
   <script src="//cdn.jsdelivr.net/npm/vue/dist/vue.min.js"></script>
-  <script src="//cdn.jsdelivr.net/npm/vuep-magic-patch@1.0.3/dist/vuep.min.js"></script>
+  <script src="//cdn.jsdelivr.net/npm/vuep-magic-patch@1.0.8/dist/vuep.min.js"></script>
   <script src="//cdn.jsdelivr.net/npm/echarts@latest/dist/echarts.min.js"></script>
   <script src="//cdn.jsdelivr.net/npm/echarts-amap/dist/echarts-amap.min.js"></script>
   <script src="//cdn.jsdelivr.net/npm/element-ui@1.4.0/lib/index.js"></script>
