+++ docs/start.md
@@ -25,28 +25,28 @@ new Vue({
 v-charts的每种图表组件，都单独打包到lib文件夹下
 ```
 |- lib/
-    |- line.js  -------------- 折线图
-    |- bar.js  --------------- 条形图
-    |- histogram.js  --------- 柱状图
-    |- pie.js  --------------- 饼图
-    |- ring.js  -------------- 环图
-    |- funnel.js  ------------ 漏斗图
-    |- waterfall.js  --------- 瀑布图
-    |- radar.js  ------------- 雷达图
-    |- map.js  --------------- 地图
-    |- sankey.js  ------------ 桑基图
-    |- heatmap.js  ----------- 热力图
-    |- scatter.js  ----------- 散点图
-    |- candle.js  ------------ k线图
-    |- gauge.js  ------------- 仪表盘
-    |- tree.js  -------------- 树图
-    |- bmap.js  -------------- 百度地图
-    |- amap.js  -------------- 高德地图
+    |- line.common.js  -------------- 折线图
+    |- bar.common.js  --------------- 条形图
+    |- histogram.common.js  --------- 柱状图
+    |- pie.common.js  --------------- 饼图
+    |- ring.common.js  -------------- 环图
+    |- funnel.common.js  ------------ 漏斗图
+    |- waterfall.common.js  --------- 瀑布图
+    |- radar.common.js  ------------- 雷达图
+    |- map.common.js  --------------- 地图
+    |- sankey.common.js  ------------ 桑基图
+    |- heatmap.common.js  ----------- 热力图
+    |- scatter.common.js  ----------- 散点图
+    |- candle.common.js  ------------ k线图
+    |- gauge.common.js  ------------- 仪表盘
+    |- tree.common.js  -------------- 树图
+    |- bmap.common.js  -------------- 百度地图
+    |- amap.common.js  -------------- 高德地图
 ```
 使用时，可以直接将单个图表引入到项目中
 ```js
 import Vue from 'vue'
-import VeLine from 'v-charts/lib/line'
+import VeLine from 'v-charts/lib/line.common'
 import App from './App.vue'
 
 Vue.component(VeLine.name, VeLine)
