+++ package.json
@@ -6,6 +6,7 @@
   "scripts": {
     "dev": "webpack-dev-server --inline --progress --config ./build/webpack.config.js",
     "build": "rm -f -r lib && node build/rollup.config.js && cp -f lib/index.min.js lib/style.min.css docs/",
+    "buildOnly": "rm -f -r lib && node build/rollup.config.js",
     "prepublishOnly": "npm run build",
     "test": "karma start ./test/karma.conf.js",
     "docs": "docsify serve docs",
@@ -30,7 +31,9 @@
     "vue": ">2.0.0"
   },
   "dependencies": {
-    "echarts-amap": "1.0.0-rc.6"
+    "echarts-amap": "1.0.0-rc.6",
+    "numerify": "1.2.8",
+    "utils-lite": "^0.1.3"
   },
   "devDependencies": {
     "autoprefixer": "^8.5.0",
@@ -69,7 +72,6 @@
     "karma-webpack": "^3.0.0",
     "less": "^3.0.4",
     "less-loader": "^4.1.0",
-    "numerify": "^1.2.8",
     "phantomjs-prebuilt": "^2.1.16",
     "prismjs": "^1.14.0",
     "rollup": "^0.60.1",
@@ -81,7 +83,6 @@
     "rollup-plugin-vue": "2.5.2",
     "style-loader": "^0.21.0",
     "uglify-es": "^3.3.9",
-    "utils-lite": "^0.1.3",
     "vue": "^2.5.16",
     "vue-loader": "^15.2.4",
     "vue-router": "^3.0.1",
