+++ src/component-list.js
@@ -1,3 +1,4 @@
+/* eslint-disable comma-dangle */
 module.exports = {
   VeBar: { src: 'src/packages/bar/index.js', dist: 'lib/bar' },
   VeHistogram: { src: 'src/packages/histogram/index.js', dist: 'lib/histogram' },
@@ -17,5 +18,5 @@ module.exports = {
   VeScatter: { src: 'src/packages/scatter/index.js', dist: 'lib/scatter' },
   VeCandle: { src: 'src/packages/candle/index.js', dist: 'lib/candle' },
   VeGauge: { src: 'src/packages/gauge/index.js', dist: 'lib/gauge' },
-  VeTree: { src: 'src/packages/tree/index.js', dist: 'lib/tree' }
+  VeTree: { src: 'src/packages/tree/index.js', dist: 'lib/tree' },
 }
