+++ src/packages/bar/index.js
@@ -1,15 +1,10 @@
-import echarts from '../../echarts-base'
 import 'echarts/lib/chart/bar'
 import { bar } from './main'
 import Core from '../../core'
-import numerify from 'numerify/lib/index.es'
-
-export default {
+export default Object.assign({}, Core, {
   name: 'VeBar',
-  mixins: [Core],
-  created () {
+  data () {
     this.chartHandler = bar
-    this.echartsLib = echarts
-  },
-  _numerify: numerify
-}
+    return {}
+  }
+})
