+++ src/packages/bmap/index.js
@@ -1,14 +1,10 @@
-import echarts from '../../echarts-base'
 import 'echarts/extension/bmap/bmap'
 import { bmap } from './main'
 import Core from '../../core'
-import numerify from 'numerify/lib/index.es'
-export default {
+export default Object.assign({}, Core, {
   name: 'VeBmap',
-  mixins: [Core],
-  created () {
+  data () {
     this.chartHandler = bmap
-    this.echartsLib = echarts
-  },
-  _numerify: numerify
-}
+    return {}
+  }
+})
