+++ src/packages/funnel/main.js
@@ -1,4 +1,4 @@
-import { itemPoint } from '../../echarts-base'
+import { itemPoint } from '../../constants'
 import { getFormated } from '../../utils'
 
 function getFunnelTooltip (dataType, digit) {
