+++ src/packages/heatmap/main.js
@@ -1,5 +1,6 @@
-import { default as echarts, itemPoint, HEAT_MAP_COLOR, HEAT_BMAP_COLOR } from '../../echarts-base'
+import { itemPoint, HEAT_MAP_COLOR, HEAT_BMAP_COLOR } from '../../constants'
 import { getBmap, getAmap, getMapJSON, getFormated } from '../../utils'
+import echarts from 'echarts/lib/echarts'
 
 function getAxisList (rows, label) {
   const result = []
