+++ src/packages/histogram/index.js
@@ -1,14 +1,10 @@
-import echarts from '../../echarts-base'
 import 'echarts/lib/chart/bar'
 import { histogram } from '../bar/main'
 import Core from '../../core'
-import numerify from 'numerify/lib/index.es'
-export default {
+export default Object.assign({}, Core, {
   name: 'VeHistogram',
-  mixins: [Core],
-  created () {
+  data () {
     this.chartHandler = histogram
-    this.echartsLib = echarts
-  },
-  _numerify: numerify
-}
+    return {}
+  }
+})
