+++ src/packages/pie/index.js
@@ -1,14 +1,10 @@
-import echarts from '../../echarts-base'
 import 'echarts/lib/chart/pie'
 import { pie } from './main'
 import Core from '../../core'
-import numerify from 'numerify/lib/index.es'
-export default {
+export default Object.assign({}, Core, {
   name: 'VePie',
-  mixins: [Core],
-  created () {
+  data () {
     this.chartHandler = pie
-    this.echartsLib = echarts
-  },
-  _numerify: numerify
-}
+    return {}
+  }
+})
