+++ src/packages/sankey/index.js
@@ -1,14 +1,10 @@
-import echarts from '../../echarts-base'
 import 'echarts/lib/chart/sankey'
 import { sankey } from './main'
 import Core from '../../core'
-import numerify from 'numerify/lib/index.es'
-export default {
+export default Object.assign({}, Core, {
   name: 'VeSankey',
-  mixins: [Core],
-  created () {
+  data () {
     this.chartHandler = sankey
-    this.echartsLib = echarts
-  },
-  _numerify: numerify
-}
+    return {}
+  }
+})
