+++ test/load/webpack/gauge/App.vue
