+++ test/load/webpack/gauge/index.js
